- various fixes of warnings
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1392 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/auth_diameter/diameter_msg.h
@@ -146,7 +146,7 @@ typedef enum {
 	AAA_APP_NASREQ               = 1,
 	AAA_APP_MOBILE_IP            = 2,
 	AAA_APP_DIAMETER_BASE_ACC    = 3,
-	AAA_APP_RELAY                = 0xffffffff,
+	AAA_APP_RELAY                = (int)0xffffffff,
 }AAA_APP_IDS;
